Split ConfigurationTests.swift file in multiple files
